[#86] moving loss functions to own file
